move oidc flow
make alias

improve docs

remove oidc
